% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivalplotqq.R
\name{plotqq.survival}
\alias{plotqq.survival}
\title{Plot Q-Q plot on survival data}
\usage{
plotqq.survival(data)
}
\arguments{
\item{data}{Data Frame}
}
\value{

}
\description{
Draw 8 different Q-Q plot Create a new factor from two existing factors, where the new factor's levels
are the union of the levels of the input factors.
}
