% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivalplotqq.R
\name{plotqq.survival}
\alias{plotqq.survival}
\title{Plot Q-Q plot for survival data}
\usage{
plotqq.survival(data)
}
\arguments{
\item{data}{Data Frame with 3 variables such as x(variable of interest), tag(binary 0 or 1, survival or not) and lnx(log x)}
}
\value{

}
\description{
Draw 8 different Q-Q plots to inference which distribution is the best to fit the survival data.
List of Candidate distributions: Weibull, Log-Normal, Gompertz, Gumbel, Normal, Pareto, Exponential, Laplace.
}
